src/userdata/requirements.txt
Django==4.2.3
psycopg2-binary
gunicorn==21.2.0
django-environ
djangorestframework
djangorestframework-simplejwt
django-otp
django-two-factor-auth[phonenumbers]
python-decouple
requests
Pillow
